v5.3.9.71 - Performance improvements and style transfer to HTML
For my small site, using a CSS file is considerable overkill. It doesn't contain many things, so there is no need to cache it and the code is not that large, so it's possible to move it to HTML, also the page should load faster as it doesn't need to download another file when opening. But if any problems appear we will switch back to the old system.